+++ lldb/source/API/SBModule.cpp
@@ -9,6 +9,7 @@
 #include "lldb/API/SBModule.h"
 #include "lldb/API/SBAddress.h"
 #include "lldb/API/SBFileSpec.h"
+#include "lldb/API/SBLineSpec.h"
 #include "lldb/API/SBModuleSpec.h"
 #include "lldb/API/SBProcess.h"
 #include "lldb/API/SBStream.h"
@@ -291,22 +292,21 @@ SBSymbolContextList SBModule::FindCompileUnits(const SBFileSpec &sb_file_spec) {
 }
 
 SBSymbolContextList
-SBModule::FindSymbolContexts(const SBLineEntry &line_entry,
-                             bool check_inlines) {
-  LLDB_INSTRUMENT_VA(this, line_entry, check_inlines);
+SBModule::FindSymbolContexts(const SBLineSpec &line_spec) {
+  LLDB_INSTRUMENT_VA(this, line_spec);
 
   SBSymbolContextList sc_list;
   const ModuleSP module_sp(GetSP());
-  if (!module_sp || !line_entry.IsValid())
+  if (!module_sp || !line_spec.IsValid())
     return sc_list;
 
-  SBFileSpec file_spec = line_entry.GetFileSpec();
+  SBFileSpec file_spec = line_spec.GetFileSpec();
   if (!file_spec.IsValid())
     return sc_list;
 
   // TODO: Pass the column when ResolveSymbolContextsForFileSpec supports it.
   module_sp->ResolveSymbolContextsForFileSpec(
-      *file_spec, line_entry.GetLine(), check_inlines,
+      *file_spec, line_spec.GetLine(), line_spec.GetCheckInlines(),
       lldb::eSymbolContextEverything, *sc_list);
   return sc_list;
 }